Nuevo componente y estilos nav
