+++ src/Components/CreateHotel/Styles.css
@@ -1,0 +1,14 @@
+/* .cardHotels-container {
+  display: flex ;
+  flex-direction: column;
+  align-items: center;
+  max-width: 1200px;
+  margin: 0 auto;
+
+}
+.form-control{
+  margin-top: 8px;
+  margin-bottom: 10px;
+  border-color: rgb(166, 219, 219);
+  width: 300px;
+  } */
